Agregado inciso d y f

+++ JPAIntegrador2/src/imodel/IStudent.java
@@ -10,6 +10,6 @@ import model.Student;
 public interface IStudent {
 	public void studentPersistence(EntityManager em, CSVParser parserStudent);
 	public void insertStudent(EntityManager em, long DNI, String name, String lastname, int age, String gender, int LU, String city);
-	public Student getStudentByLU(EntityManager em, int LU);
+	public Student getStudentByLU(EntityManager em, Long LU);
 	public List<Student> getStudentsByGender(EntityManager em, String gender);
 }
